Tested all review routes, added example responses

+++ server/app.py
@@ -268,30 +268,99 @@ api.add_resource(RestaurantsById, "/restaurants/<int:id>")
 #post
 class Reviews(Resource):
 
-    def get(self):
-        reviews = Review.query.all()
-        reviews_list = [review.to_dict for review in reviews]
+    def get(self, restaurant_id):
+        restaurant = Restaurant.query.filter_by(id=restaurant_id).first()
+        reviews_list = [review.to_dict(rules=("-restaurant.reservations",)) for review in restaurant.reviews]
 
         return make_response(reviews_list, 200)
-
-    def post(self):
+    
+    #Single review in a list of all reviews for a specific restaurant.  Response in format:
+    
+    '''
+    [
+    {
+        "comment": "The food was amazing!",
+        "id": 1,
+        "rating": 5,
+        "restaurant": {
+            "address": "550 3rd Avenue",
+            "capacity": 60,
+            "category": "Chinese",
+            "city": "New York",
+            "close_time": 2000,
+            "id": 1,
+            "image": "https://assets-global.website-files.com/645aecde0bd10564f39f4379/64e73f688681676ffdf401d0_20230103%20Little%20Alley_367%20WHD.jpg",
+            "menu_link": "https://www.littlealley.nyc/menu",
+            "name": "Little Alley",
+            "open_time": 1600,
+            "phone": "646-998-3976",
+            "res_duration": 90,
+            "state": "NY",
+            "website": "https://www.littlealley.nyc/",
+            "zip": "10016"
+        },
+        "restaurant_id": 1,
+        "timestamp": 1735012800,
+        "user": {
+            "IsAdmin": false,
+            "email": "john@example.com",
+            "first_name": "John",
+            "id": 1,
+            "last_name": "Doe",
+            "password": "password123",
+            "phone": "123-456-7890",
+            "username": "johndoe"
+        },
+        "user_id": 1
+    },
+    '''
+    
+    def post(self, restaurant_id):
         data = request.get_json()
+        
+        restaurant = Restaurant.query.get(restaurant_id)
+        
+        if not restaurant:
+            return make_response({'error': 'Restaurant not found'}, 404)
 
         try:
             review = Review(
                 timestamp=data.get("timestamp"),
                 rating=data.get("rating"),
                 comment=data.get("comment"),
+                user_id = data.get("user_id"),
+                restaurant_id = restaurant_id
             )
             db.session.add(review)
             db.session.commit()
-            return make_response(review.to_dict(), 201)
+            return make_response(review.to_dict(rules =("-restaurant", )), 201)
 
         except ValueError:
             return make_response({"errors": ["validation errors"]}, 400)
-
-
-api.add_resource(Reviews, "/reviews")
+        
+    #Posting a review to a specific restaurant. Response:
+    '''
+    {
+        "comment": "ISHHH WHO EATS CAT FOOD",
+        "id": 84,
+        "rating": 1,
+        "restaurant_id": 1,
+        "timestamp": "CHANGED",
+        "user": {
+            "IsAdmin": false,
+            "email": "john@example.com",
+            "first_name": "John",
+            "id": 1,
+            "last_name": "Doe",
+            "password": "password123",
+            "phone": "123-456-7890",
+            "username": "johndoe"
+        },
+        "user_id": 1
+    }
+    '''
+    
+api.add_resource(Reviews, "/restaurants/<int:restaurant_id>/reviews")
 
 
 #ReviewsById Routes
@@ -299,29 +368,65 @@ api.add_resource(Reviews, "/reviews")
 #delete
 class ReviewsById(Resource):
 
-    def patch(self, id):
-        review = Review.query.filter_by(id=id).first()
+    def patch(self, restaurant_id, review_id):
+        restaurant = Restaurant.query.get(restaurant_id)
 
+        if not restaurant:
+            return make_response({"error": "Restaurant not found"}, 404)
+        
+        review = Review.query.filter_by(restaurant_id=restaurant_id, id = review_id).first()
+        
         if not review:
             return make_response({"error": "Review not found"}, 404)
 
         try:
             data = request.get_json()
-            for attr in data:
-                setattr(review, attr, data.get(attr))
+            # for attr in data:
+            #     setattr(restaurant.review, attr, data.get(attr))
+            for key, value in data.items():
+                if hasattr(review, key):
+                    setattr(review, key, value)
+            
 
             db.session.add(review)
             db.session.commit()
-            return make_response(review.to_dict(), 202)
+            return make_response(review.to_dict(rules=("-restaurant",)), 202)
 
         except ValueError:
             return ({"errors": ["validation errors"]}, 400)
+        
+    '''
+    {
+        "comment": "YUM CAT FOOD!",
+        "id": 85,
+        "rating": 1,
+        "restaurant_id": 1,
+        "timestamp": "CHANGED",
+        "user": {
+            "IsAdmin": false,
+            "email": "john@example.com",
+            "first_name": "John",
+            "id": 1,
+            "last_name": "Doe",
+            "password": "password123",
+            "phone": "123-456-7890",
+            "username": "johndoe"
+        },
+        "user_id": 1
+    }
+    '''
+        
 
-    def delete(self, id):
-        review = Review.query.filter_by(id=id).first()
+    def delete(self, restaurant_id, review_id):
+        restaurant = Restaurant.query.get(restaurant_id)
 
+        if not restaurant:
+            return make_response({"error": "Restaurant not found"}, 404)
+        
+        review = Review.query.filter_by(restaurant_id=restaurant_id, id = review_id).first()
+        
         if not review:
-            return make_response({"error": "review not found"}, 404)
+            return make_response({"error": "Review not found"}, 404)
 
         try:
             db.session.delete(review)
@@ -331,9 +436,14 @@ class ReviewsById(Resource):
         except Exception as e:
             print(f"Error deleting review: {str(e)}")
             return ({"errors": ["validation errors"]}, 400)
+        
+        
+    '''
+    Empty response 204, tested and persists
+    '''
 
 
-api.add_resource(ReviewsById, "/reviews/<int:id>")
+api.add_resource(ReviewsById, "/restaurants/<int:restaurant_id>/reviews/<int:review_id>")
 
 
 
@@ -350,7 +460,7 @@ class Reservations(Resource):
 
     def get(self):
         reservations = Reservation.query.all()
-        reservations_list = [reservation.to_dict for reservation in reservations]
+        reservations_list = [reservation.to_dict() for reservation in reservations]
 
         return make_response(reservations_list, 200)
 
@@ -432,7 +542,7 @@ class Users(Resource):
 
     def get(self):
         users = User.query.all()
-        users_list = [user.to_dict for user in users]
+        users_list = [user.to_dict() for user in users]
 
         return make_response(users_list, 200)
